Mac line endings baffled VS8.

+++ sf-cvs/trunk/lib-src/soundtouch/source/SoundTouch/FIRFilter.cpp
@@ -1,8 +1,8 @@
 ////////////////////////////////////////////////////////////////////////////////
 ///
-/// General FIR digital filter routines with MMX optimization. 
+/// General FIR digital filter routines with MMX optimization.
 ///
-/// Note : MMX optimized functions reside in a separate, platform-specific file, 
+/// Note : MMX optimized functions reside in a separate, platform-specific file,
 /// e.g. 'mmx_win.cpp' or 'mmx_gcc.cpp'
 ///
 /// Author        : Copyright (c) Olli Parviainen
@@ -11,10 +11,10 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////
 //
-// Last changed  : $Date: 2006-09-18 07:31:40 $
-// File revision : $Revision: 1.3 $
+// Last changed  : $Date: 2006-09-18 22:29:22 $
+// File revision : $Revision: 1.4 $
 //
-// $Id: FIRFilter.cpp,v 1.3 2006-09-18 07:31:40 richardash1981 Exp $
+// $Id: FIRFilter.cpp,v 1.4 2006-09-18 22:29:22 martynshaw Exp $
 //
 ////////////////////////////////////////////////////////////////////////////////
 //
@@ -84,14 +84,14 @@ uint FIRFilter::evaluateFilterStereo(SAMPLETYPE *dest, const SAMPLETYPE *src, ui
 
     end = 2 * (numSamples - length);
 
-    for (j = 0; j < end; j += 2) 
+    for (j = 0; j < end; j += 2)
     {
         const SAMPLETYPE *ptr;
 
         suml = sumr = 0;
         ptr = src + j;
 
-        for (i = 0; i < length; i += 4) 
+        for (i = 0; i < length; i += 4)
         {
             // loop is unrolled by factor of 4 here for efficiency
             suml += ptr[2 * i + 0] * filterCoeffs[i + 0] +
@@ -139,15 +139,15 @@ uint FIRFilter::evaluateFilterMono(SAMPLETYPE *dest, const SAMPLETYPE *src, uint
     assert(length != 0);
 
     end = numSamples - length;
-    for (j = 0; j < end; j ++) 
+    for (j = 0; j < end; j ++)
     {
         sum = 0;
-        for (i = 0; i < length; i += 4) 
+        for (i = 0; i < length; i += 4)
         {
             // loop is unrolled by factor of 4 here for efficiency
-            sum += src[i + 0] * filterCoeffs[i + 0] + 
-                   src[i + 1] * filterCoeffs[i + 1] + 
-                   src[i + 2] * filterCoeffs[i + 2] + 
+            sum += src[i + 0] * filterCoeffs[i + 0] +
+                   src[i + 1] * filterCoeffs[i + 1] +
+                   src[i + 2] * filterCoeffs[i + 2] +
                    src[i + 3] * filterCoeffs[i + 3];
         }
 #ifdef INTEGER_SAMPLES
@@ -177,7 +177,7 @@ void FIRFilter::setCoefficients(const SAMPLETYPE *coeffs, uint newLength, uint u
     assert(length == newLength);
 
     resultDivFactor = uResultDivFactor;
-    resultDivider = (SAMPLETYPE)pow(2, resultDivFactor);
+    resultDivider = (SAMPLETYPE)pow(2., (double)resultDivFactor);
 
     delete[] filterCoeffs;
     filterCoeffs = new SAMPLETYPE[length];
@@ -192,9 +192,9 @@ uint FIRFilter::getLength() const
 
 
 
-// Applies the filter to the given sequence of samples. 
+// Applies the filter to the given sequence of samples.
 //
-// Note : The amount of outputted samples is by value of 'filter_length' 
+// Note : The amount of outputted samples is by value of 'filter_length'
 // smaller than the amount of input samples.
 uint FIRFilter::evaluate(SAMPLETYPE *dest, const SAMPLETYPE *src, uint numSamples, uint numChannels) const
 {
@@ -204,7 +204,7 @@ uint FIRFilter::evaluate(SAMPLETYPE *dest, const SAMPLETYPE *src, uint numSample
     assert(lengthDiv8 * 8 == length);
     if (numSamples < length) return 0;
     assert(resultDivFactor >= 0);
-    if (numChannels == 2) 
+    if (numChannels == 2)
     {
         return evaluateFilterStereo(dest, src, numSamples);
     } else {
@@ -214,7 +214,7 @@ uint FIRFilter::evaluate(SAMPLETYPE *dest, const SAMPLETYPE *src, uint numSample
 
 
 
-// Operator 'new' is overloaded so that it automatically creates a suitable instance 
+// Operator 'new' is overloaded so that it automatically creates a suitable instance
 // depending on if we've a MMX-capable CPU available or not.
 void * FIRFilter::operator new(size_t s)
 {
